requirements.txt
loguru
python-dotenv
rich
pyinstaller